requirements.txt
python-telegram-bot==20.3
opencv-python
pynput
Pillow

